requirements.txt
Plaintext
streamlit
google-generativeai